api/requirements.txt
fastapi
uvicorn
onnxruntime
numpy
pillow
python-multipart
starlette